Migrate form functionality into form.js
